requirements/web.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
aiofiles
langgraph
langchain
langchain-anthropic
pydantic
python-dotenv
pyyaml
pdfplumber
python-docx
pytesseract
pillow
requests
rich